:recycle:  Refactor: 컴포넌트 분리 -> FirstCalculatorForm 생성
Closes #16
