+++ src/components/domain/index.jsx
@@ -1,1 +1,2 @@
 export { default as FirstCalculator } from './FirstCalculator'
+export { default as FirstCalculatorForm } from './FirstCalculatorForm'
